pyproject.toml
[project]
name = "beancount-ledger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.2"
dependencies = [
    "beancount>=3.2.0",
    "beangulp>=0.2.0",
    "fava>=1.30.10",
    "jinja2>=3.1.6",
    "pandas>=2.3.3",
    "pdf2image>=1.17.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "weasyprint>=67.0",
]
